Update intro to adhere an Amazon cert policy.
The skilled failed the certification due to: 'Skills which reference,
replicate, or otherwise create confusion between the Alexa first-party
personality and the third-party skill are not appropriate for Alexa
Skills Catalog.' The intro was updated to remove Alice as "Alexa's twin
sister". In the future, SSML/Polly can be used to further remove
confusion between Alexa and Alice.